extra work for toplevelwindow
