+++ Authwindow.py
@@ -1,36 +1,44 @@
+from tkinter import StringVar
+
 import customtkinter
 import os
 import subprocess
 
 from api import entry
-from sub.exchange_code_for_token import *
+from sub.auth_code import *
 
 
 class ToplevelWindow(customtkinter.CTkToplevel):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.geometry("400x300")
+        self.geometry("200x140")
+        self.minsize(200, 140)
 
+        #---- Check Auth File and Open Browser
         if not os.path.exists('.env'):
             with open('.env', 'w') as f:
                 pass
         get_authorization_code()
+
+        #---- Start Top Level Window
         self.title("Authorization Code")
 
         self.label = customtkinter.CTkLabel(self, text="Enter Authorization Code:")
-        self.label.pack(padx=20, pady=20)
+        self.label.pack(padx=20, pady=5)
 
-        entry = customtkinter.CTkEntry(self)
-        print(entry.get())
-        entry.pack()
+        self.wronglabel = customtkinter.CTkLabel(self, text="")
+        self.wronglabel.pack()
+
+        self.authkey = StringVar()
+        self.entry = customtkinter.CTkEntry(self, textvariable=self.authkey)
+        self.entry.pack()
 
         self.submit_button = customtkinter.CTkButton(self, text="Submit", command=self.submit_authorization_code)
-        self.submit_button.pack()
+        self.submit_button.pack(pady="10")
 
     def submit_authorization_code(self):
-        print("1")
-        print(entry.get())
-        authorization_code = entry.get()
+        print(self.authkey.get())
+        authorization_code = self.authkey.get()
         if authorization_code:
             access_token = exchange_code_for_token(authorization_code.strip())
             if access_token:
@@ -39,7 +47,5 @@ class ToplevelWindow(customtkinter.CTkToplevel):
                 self.destroy()
             else:
                 print("Authorization failed.")
-            self.destroy()
         else:
-            print("Please enter the authorization code.")
-
+            self.wronglabel.configure(text="Paste Authentication Code", text_color="red")
